// backend/src/types/auth.types.ts
export interface LoginRequest {
    email: string;
    password: string;
}   

export interface AuthUser {
    id: number;
    email: string;
    name: string;
    role: Role;
    createdAt: Date;
}

export interface LoginResponse {
    user: AuthUser;
    token: string;
}

export enum Role {
  ADMIN = "ADMIN",
  USER = "USER",
}